Volume: Minor type correction (float-->integer) required for the Numeric version I'm using on Windows.

+++ src/cclib/method/volume.py
@@ -37,7 +37,7 @@ class Volume(object):
         self.topcorner = topcorner
         self.numpts = []
         for i in range(3):
-            self.numpts.append( (self.topcorner[i]-self.origin[i])/self.spacing[i] + 1 )
+            self.numpts.append(int((self.topcorner[i]-self.origin[i])/self.spacing[i]                                   + 1) )
         self.data = Numeric.zeros( tuple(self.numpts), "d")
 
     def __str__(self):
